backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
Pillow==9.0.0
numpy==1.21.0
scikit-learn==1.0.2 